pyproject.toml
[tool.poetry]
name = "politech-parser"
version = "0.1.0"
description = ""
authors = ["Shmonov Matvey <meology77@gmail.com>"]
readme = "README.md"
packages = [{include = "politech_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
aiohttp-retry = "^2.8.3"
aiohttp-socks = "^0.8.0"
python-dotenv = "^1.0.0"
aiogram = {version = "^3.0.0b7", allow-prereleases = true}


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
isort = "^5.12.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
